local-qa-bot/requirements.txt
langchain
openai
faiss-cpu
python-dotenv
tqdm